refactor: update import paths for S3 plugin to use amapy_plugin_s3
