+++ client/asset-manager/asset_manager/plugins.py
@@ -4,7 +4,7 @@ from amapy_pluggy import register_plugins as pluggy
 from amapy_plugin_gcr import GcrStoragePlugin
 from amapy_plugin_gcs import GcsStoragePlugin
 from asset_plugin_posix import PosixStoragePlugin
-from asset_plugin_s3 import AwsStoragePlugin
+from amapy_plugin_s3 import AwsStoragePlugin
 
 BUNDLED_PLUGINS = [
     GcsStoragePlugin,
