+++ client/asset-plugin-s3/amapy_plugin_s3/aws_http_handler.py
@@ -8,9 +8,9 @@ from botocore.errorfactory import ClientError
 
 from amapy_pluggy.storage import StorageData, StorageURL, BlobStoreURL
 from amapy_pluggy.storage import storage_utils
-from asset_plugin_s3.aws_auth import AwsAuth, get_aws_id_k_date
-from asset_plugin_s3.aws_blob import AwsBlob
-from asset_plugin_s3.transporter import AsyncAwsTransporter
+from amapy_plugin_s3.aws_auth import AwsAuth, get_aws_id_k_date
+from amapy_plugin_s3.aws_blob import AwsBlob
+from amapy_plugin_s3.transporter import AsyncAwsTransporter
 from amapy_utils.utils import utils
