+++ client/asset-plugin-s3/amapy_plugin_s3/aws_storage_mixin.py
@@ -2,8 +2,8 @@ from functools import cached_property
 from typing import Union
 
 from amapy_pluggy.storage import StorageData, StorageURL
-from asset_plugin_s3.aws_http_handler import AwsHttpHandler
-from asset_plugin_s3.aws_mount_handler import AwsMountHandler
+from amapy_plugin_s3.aws_http_handler import AwsHttpHandler
+from amapy_plugin_s3.aws_mount_handler import AwsMountHandler
 
 
 class AwsStorageMixin:
