+++ client/asset-plugin-s3/amapy_plugin_s3/bucket_cors.py
