+++ client/asset-plugin-s3/amapy_plugin_s3/mounted_bucket/async_mount_copy.py
@@ -4,7 +4,7 @@ import os
 import aiofiles
 import backoff
 
-from asset_plugin_s3.mounted_bucket.mounted_resource import AwsMountedResource
+from amapy_plugin_s3.mounted_bucket.mounted_resource import AwsMountedResource
 from amapy_utils.utils.log_utils import get_logger
 
 logger = get_logger(__name__)
