+++ client/asset-plugin-s3/amapy_plugin_s3/mounted_bucket/async_mount_transporter.py
@@ -1,27 +1,37 @@
 from botocore.exceptions import ClientError
 
 from amapy_pluggy.storage.transporter import Transporter, TransportResource
-from asset_plugin_s3.transporter.async_aws import async_upload, async_download, async_copy, async_update_blob
-from asset_plugin_s3.transporter.aws_transport_resource import AwsUploadResource, AwsDownloadResource, AwsCopyResource
+from amapy_plugin_s3.mounted_bucket import async_mount_copy
+from amapy_plugin_s3.mounted_bucket.mounted_resource import AwsMountedResource
 from amapy_utils.common import exceptions
 from amapy_utils.utils import utils
 
 
-class AsyncAwsTransporter(Transporter):
+class AsyncMountTransporter(Transporter):
+    mount_cfg = None
+
+    @classmethod
+    def shared(cls, mount_cfg=None, **kwargs):
+        if not hasattr(cls, 'instance'):
+            instance = super(AsyncMountTransporter, cls).shared(**kwargs)
+            instance.mount_cfg = mount_cfg
+            cls.instance = instance
+        return cls.instance
 
     def get_download_resource(self, src: str, dst: str, src_hash: tuple) -> TransportResource:
-        return AwsDownloadResource(src=src, dst=dst, hash=src_hash)
+        return AwsMountedResource(src=src, dst=dst, src_hash=src_hash, mount_cfg=self.mount_cfg)
 
     def get_upload_resource(self, src: str, dst: str, src_hash: tuple) -> TransportResource:
-        return AwsUploadResource(src=src, dst=dst, hash=src_hash)
+        return AwsMountedResource(src=src, dst=dst, src_hash=src_hash, mount_cfg=self.mount_cfg)
 
     def get_copy_resource(self, src: str, dst: str, src_hash: tuple, **kwargs) -> TransportResource:
-        return AwsCopyResource(src=src, dst=dst, hash=src_hash, **kwargs)
+        return AwsMountedResource(src=src, dst=dst, src_hash=src_hash, mount_cfg=self.mount_cfg, **kwargs)
 
     def upload(self, resources: [TransportResource]) -> None:
         try:
             for chunk in utils.batch(resources, batch_size=self.batch_size):
-                async_upload.upload_resources(credentials=self.credentials, resources=chunk)
+                async_mount_copy.copy_resources(resources=chunk)
+        # keep the exception handling as is, since fuse-mounting uses http under the hood
         except ClientError as e:
             if e.response["Error"]["Code"] == "SignatureDoesNotMatch":
                 raise exceptions.InvalidStorageCredentialsError("Credentials expired. Fetch Again.") from e
@@ -34,35 +44,17 @@ class AsyncAwsTransporter(Transporter):
     def download(self, resources: [TransportResource]) -> None:
         try:
             for chunk in utils.batch(resources, batch_size=self.batch_size):
-                async_download.download_resources(credentials=self.credentials, resources=chunk)
+                async_mount_copy.copy_resources(resources=chunk)
         except Exception as e:
             if isinstance(e, ClientError) and e.response["Error"]["Code"] == "SignatureDoesNotMatch":
                 raise exceptions.InvalidStorageCredentialsError("Credentials expired. Fetch Again.") from e
             raise exceptions.AssetException("Error while downloading resources") from e
 
     def copy(self, resources: [TransportResource]) -> None:
-        # update multipart sizes of resource if it has a blob
-        self.update_multipart_sizes(resources=resources)
         try:
             for chunk in utils.batch(resources, batch_size=self.batch_size):
-                async_copy.copy_resources(credentials=self.credentials, resources=chunk)
+                async_mount_copy.copy_resources(resources=chunk)
         except Exception as e:
             if isinstance(e, ClientError) and e.response["Error"]["Code"] == "SignatureDoesNotMatch":
                 raise exceptions.InvalidStorageCredentialsError("Credentials expired. Fetch Again.") from e
             raise exceptions.AssetException("Error while copying resources") from e
-
-    def update_multipart_sizes(self, resources: [AwsCopyResource]) -> None:
-        blobs = [resource.blob for resource in resources if hasattr(resource, "blob") and resource.blob.is_multipart]
-        if not blobs:
-            return
-        # fetch and update multipart sizes of blobs
-        self.update_multipart_blobs(blobs=blobs)
-
-    def update_multipart_blobs(self, blobs: list) -> None:
-        try:
-            for chunk in utils.batch(blobs, batch_size=self.batch_size):
-                async_update_blob.update_multipart_sizes(credentials=self.credentials, blobs=chunk)
-        except Exception as e:
-            if isinstance(e, ClientError) and e.response["Error"]["Code"] == "SignatureDoesNotMatch":
-                raise exceptions.InvalidStorageCredentialsError("Credentials expired. Fetch Again.") from e
-            raise exceptions.AssetException("Error while fetching blob multipart size") from e
