+++ client/asset-plugin-s3/amapy_plugin_s3/mounted_bucket/mounted_url.py
