+++ client/asset-plugin-s3/amapy_plugin_s3/mounted_bucket/tests/test_mounted_url.py
@@ -2,7 +2,7 @@ import pytest
 
 from amapy_pluggy.storage import BlobStoreURL
 from amapy_pluggy.storage.mount_config import MountConfig
-from asset_plugin_s3.mounted_bucket.mounted_url import MountedBlobStoreURL
+from amapy_plugin_s3.mounted_bucket.mounted_url import MountedBlobStoreURL
 from amapy_utils.common import exceptions
 
 
@@ -19,7 +19,7 @@ def mount_config():
 
 @pytest.fixture
 def mock_blob_store_url(monkeypatch):
-    monkeypatch.setattr('asset_plugin_s3.mounted_bucket.mounted_url.MountedBlobStoreURL', MockBlobStoreURL)
+    monkeypatch.setattr('amapy_plugin_s3.mounted_bucket.mounted_url.MountedBlobStoreURL', MockBlobStoreURL)
 
 
 @pytest.mark.usefixtures("mock_blob_store_url")
