+++ client/asset-plugin-s3/amapy_plugin_s3/s3_proxy.py
