+++ client/asset-plugin-s3/amapy_plugin_s3/tests/test_aws_storage.py
@@ -7,8 +7,8 @@ import pytest
 
 from amapy_pluggy.storage import BlobStoreURL
 from amapy_pluggy.storage.storage_credentials import StorageCredentials
-from asset_plugin_s3.aws_blob import AwsBlob
-from asset_plugin_s3.aws_storage import AwsStorage
+from amapy_plugin_s3.aws_blob import AwsBlob
+from amapy_plugin_s3.aws_storage import AwsStorage
 from amapy_utils.utils import time_it
