+++ client/asset-plugin-s3/amapy_plugin_s3/transporter/async_aws/async_copy.py
@@ -7,8 +7,8 @@ import backoff
 from botocore.awsrequest import AWSRequest
 from botocore.credentials import Credentials
 
-from asset_plugin_s3.aws_auth import AwsAuth
-from asset_plugin_s3.transporter.aws_transport_resource import AwsCopyResource
+from amapy_plugin_s3.aws_auth import AwsAuth
+from amapy_plugin_s3.transporter.aws_transport_resource import AwsCopyResource
 from amapy_utils.utils.log_utils import get_logger
 
 logger = get_logger(__name__)
