+++ client/asset-plugin-s3/amapy_plugin_s3/transporter/async_aws/async_upload.py
@@ -6,7 +6,7 @@ import aiohttp
 import backoff
 from aiobotocore.config import AioConfig
 
-from asset_plugin_s3.transporter.aws_transport_resource import AwsUploadResource
+from amapy_plugin_s3.transporter.aws_transport_resource import AwsUploadResource
 from amapy_utils.utils.log_utils import get_logger
 
 logger = get_logger(__name__)
