+++ client/asset-plugin-s3/amapy_plugin_s3/transporter/aws_transport_resource.py
