+++ client/asset-plugin-s3/amapy_plugin_s3/transporter/legacy_aws/async_download.py
@@ -9,7 +9,7 @@ import backoff
 from aiobotocore.session import get_session
 
 # native
-from asset_plugin_s3.transporter.aws_transport_resource import AwsDownloadResource
+from amapy_plugin_s3.transporter.aws_transport_resource import AwsDownloadResource
 from amapy_utils.utils.file_utils import FileUtils
 # plugins
 from amapy_utils.utils.log_utils import get_logger
