+++ client/asset-plugin-s3/amapy_plugin_s3/transporter/legacy_aws/async_upload.py
@@ -8,7 +8,7 @@ import backoff
 from aiobotocore.session import get_session
 
 # native
-from asset_plugin_s3.transporter.aws_transport_resource import AwsUploadResource
+from amapy_plugin_s3.transporter.aws_transport_resource import AwsUploadResource
 # plugins
 from amapy_utils.utils.log_utils import get_logger
