+++ client/asset-plugin-s3/amapy_plugin_s3/transporter/legacy_aws/legacy_aws_transporter.py
@@ -1,6 +1,6 @@
 from amapy_pluggy.storage.transporter import Transporter, TransportResource
-from asset_plugin_s3.transporter.aws_transport_resource import AwsUploadResource, AwsDownloadResource
-from asset_plugin_s3.transporter.legacy_aws import async_upload, async_download
+from amapy_plugin_s3.transporter.aws_transport_resource import AwsUploadResource, AwsDownloadResource
+from amapy_plugin_s3.transporter.legacy_aws import async_upload, async_download
 from amapy_utils.utils import utils
