+++ client/asset-plugin-s3/amapy_plugin_s3/transporter/legacy_aws/tests/test_legacy_aws_transporter.py
@@ -2,9 +2,9 @@ import os
 from datetime import datetime
 from unittest.mock import MagicMock, patch
 
-from asset_plugin_s3.aws_storage import AwsStorage
-from asset_plugin_s3.transporter.aws_transport_resource import AwsUploadResource, AwsDownloadResource
-from asset_plugin_s3.transporter.legacy_aws.legacy_aws_transporter import LegacyAwsTransporter
+from amapy_plugin_s3.aws_storage import AwsStorage
+from amapy_plugin_s3.transporter.aws_transport_resource import AwsUploadResource, AwsDownloadResource
+from amapy_plugin_s3.transporter.legacy_aws.legacy_aws_transporter import LegacyAwsTransporter
 from amapy_utils.utils import utils
 
 
@@ -27,9 +27,9 @@ def mock_async_upload_resource(mock_upload):
 
 def test_upload(project_root: str, upload_test_url: str, aws_test_credentials: dict):
     with patch('amapy_pluggy.storage.storage_credentials.StorageCredentials.shared') as mock_shared_storage, \
-            patch('asset_plugin_s3.aws_storage.AwsStorage.shared') as mock_Aws_Storage, \
+            patch('amapy_plugin_s3.aws_storage.AwsStorage.shared') as mock_Aws_Storage, \
             patch('aioboto3.Session.client') as mock_client, \
-            patch('asset_plugin_s3.transporter.legacy_aws.async_upload.async_upload_resource') as mock_upload:
+            patch('amapy_plugin_s3.transporter.legacy_aws.async_upload.async_upload_resource') as mock_upload:
         mock_s3_client = MagicMock()
         mock_client.return_value = mock_s3_client
         # Mock AwsStorage credentials
@@ -61,9 +61,9 @@ def test_upload(project_root: str, upload_test_url: str, aws_test_credentials: d
 
 def test_upload_dir(project_root: str, upload_test_url: str, aws_test_credentials: dict):
     with patch('amapy_pluggy.storage.storage_credentials.StorageCredentials.shared') as mock_shared_storage, \
-            patch('asset_plugin_s3.aws_storage.AwsStorage.shared') as mock_Aws_Storage, \
+            patch('amapy_plugin_s3.aws_storage.AwsStorage.shared') as mock_Aws_Storage, \
             patch('aioboto3.Session.client') as mock_client, \
-            patch('asset_plugin_s3.transporter.legacy_aws.async_upload.async_upload_resource') as mock_upload:
+            patch('amapy_plugin_s3.transporter.legacy_aws.async_upload.async_upload_resource') as mock_upload:
         mock_s3_client = MagicMock()
         mock_client.return_value = mock_s3_client
         # Mock AwsStorage credentials
@@ -99,10 +99,10 @@ def mock_download_resource(mock_async_download):
 
 def test_download(project_root, aws_test_credentials):
     with patch('amapy_pluggy.storage.storage_credentials.StorageCredentials.shared') as mock_shared_storage, \
-            patch('asset_plugin_s3.aws_storage.AwsStorage.shared') as mock_Aws_Storage, \
+            patch('amapy_plugin_s3.aws_storage.AwsStorage.shared') as mock_Aws_Storage, \
             patch('aioboto3.Session.client') as mock_client, \
             patch(
-                'asset_plugin_s3.transporter.legacy_aws.async_download.__async_download_resource') as mock_async_download:
+                'amapy_plugin_s3.transporter.legacy_aws.async_download.__async_download_resource') as mock_async_download:
         mock_s3_client = MagicMock()
         mock_client.return_value = mock_s3_client
         # Mock AwsStorage credentials
