--- client/asset-plugin-s3/asset_plugin_s3/transporter/__init__.py
@@ -1,2 +1,0 @@
-from asset_plugin_s3.transporter.async_aws.async_aws_transporter import AsyncAwsTransporter
-from asset_plugin_s3.transporter.legacy_aws.legacy_aws_transporter import LegacyAwsTransporter
